[PDK-9591][PDK-10839] Winbond OSPI NAND Flash Driver Development
- added support for serial NAND Flash
- added UTs for Single SPI mode, Octal DAC, INDAC and DAC+DMA modes
- PHY tuning support pending

+++ packages/ti/board/src/flash/nand/nand.h
@@ -144,7 +144,7 @@ typedef enum NAND_EccAlgo
 typedef struct
 {
     /*! NAND HW interface handle */
-    uint32_t hwHandle;
+    uintptr_t hwHandle;
     /*! manufacturer ID*/
     uint32_t manufacturerId;
     /*! Manufacturers device ID */
